clean up implementation
