+++ module/state_synchronization/indexer/indexer_core.go
@@ -8,12 +8,7 @@ import (
 	"github.com/rs/zerolog"
 	"golang.org/x/sync/errgroup"
 
-	"github.com/onflow/cadence"
-	"github.com/onflow/cadence/encoding/ccf"
-	"github.com/onflow/cadence/runtime/common"
-	"github.com/onflow/cadence/runtime/stdlib"
 	"github.com/onflow/flow-go/fvm/storage/derived"
-	"github.com/onflow/flow-go/fvm/storage/snapshot"
 	"github.com/onflow/flow-go/ledger"
 	"github.com/onflow/flow-go/ledger/common/convert"
 	"github.com/onflow/flow-go/model/flow"
@@ -341,19 +336,16 @@ func (c *IndexerCore) updateProgramCache(height uint64, events []flow.Event, col
 		header.ParentID,
 	)
 
-	// step 1:
-	// NewSnapshotReadTableTransaction()
-	// need to add special mode to allow bypassing the isScript check on line 290
+	// configure the derived transaction data to allow scripts to cache programs
 	derivedBlockData.AllowWritesInReadonlyTransaction()
 
-	// step 2:
-	// invalidate data:
-	// cycle through events and grab AccountContractUpdated, Deployed
+	// get a list of all contracts that were updated in this block
 	invalidatedPrograms, err := findContractUpdates(events)
 	if err != nil {
 		return fmt.Errorf("could not find contract updates for block %d: %w", height, err)
 	}
 
+	// invalidate cache entries for all modified programs
 	tx, err := derivedBlockData.NewDerivedTransactionData(0, 0)
 	if err != nil {
 		return fmt.Errorf("could not create derived transaction data for block %d: %w", height, err)
@@ -373,113 +365,5 @@ func (c *IndexerCore) updateProgramCache(height uint64, events []flow.Event, col
 		return fmt.Errorf("could not commit derived transaction data for block %d: %w", height, err)
 	}
 
-	// invalidator:
-	// program => invalidate programs that had changes
-	// meter => service account was authorizer
-
-	// scripts code payloads
-
-	// concurrency may be an issue
-
-	// see derived chain data tests for examples
-
 	return nil
-
-}
-
-// hasAuthorizedTransaction checks if the provided account was an authorizer in any of the transactions.
-func hasAuthorizedTransaction(collections []*flow.Collection, address flow.Address) bool {
-	for _, collection := range collections {
-		for _, tx := range collection.Transactions {
-			for _, authorizer := range tx.Authorizers {
-				if authorizer == address {
-					return true
-				}
-			}
-		}
-	}
-
-	return false
-}
-
-// findContractUpdates returns a map of common.AddressLocation for all contracts updated within the
-// given events.
-func findContractUpdates(events []flow.Event) (map[common.AddressLocation]struct{}, error) {
-	accountContractUpdated := flow.EventType(stdlib.AccountContractUpdatedEventType.ID())
-
-	invalidatedPrograms := make(map[common.AddressLocation]struct{})
-	for _, event := range events {
-		if event.Type == accountContractUpdated {
-			location, err := parseAccountContractUpdated(event)
-			if err != nil {
-				return nil, fmt.Errorf("could not parse account contract updated event: %w", err)
-			}
-			invalidatedPrograms[location] = struct{}{}
-		}
-	}
-	return invalidatedPrograms, nil
-}
-
-// parseAccountContractUpdated parses an account contract updated event and returns the address location.
-func parseAccountContractUpdated(event *flow.Event) (common.AddressLocation, error) {
-	payload, err := ccf.Decode(nil, event.Payload)
-	if err != nil {
-		return common.AddressLocation{}, fmt.Errorf("could not unmarshal event payload: %w", err)
-	}
-
-	cdcEvent, ok := payload.(cadence.Event)
-	if !ok {
-		return common.AddressLocation{}, fmt.Errorf("invalid event payload type: %T", payload)
-	}
-
-	address, ok := cdcEvent.Fields[0].(cadence.Address)
-	if !ok {
-		return common.AddressLocation{}, fmt.Errorf("invalid cadence type for address: %T", cdcEvent.Fields[0])
-	}
-
-	contractName, ok := cdcEvent.Fields[2].(cadence.String)
-	if !ok {
-		return common.AddressLocation{}, fmt.Errorf("invalid cadence type for contract name: %T", cdcEvent.Fields[2])
-	}
-
-	return common.NewAddressLocation(nil, common.Address(address), contractName.String()), nil
-}
-
-type accessInvalidator struct {
-	programs            *programInvalidator
-	meterParamOverrides *meterParamOverridesInvalidator
-}
-
-func (inv *accessInvalidator) ProgramInvalidator() derived.ProgramInvalidator {
-	return inv.programs
-}
-
-func (inv *accessInvalidator) MeterParamOverridesInvalidator() derived.MeterParamOverridesInvalidator {
-	return inv.meterParamOverrides
-}
-
-type programInvalidator struct {
-	invalidateAll bool
-	invalidated   map[common.AddressLocation]struct{}
-}
-
-func (inv *programInvalidator) ShouldInvalidateEntries() bool {
-	return inv.invalidateAll
-}
-
-func (inv *programInvalidator) ShouldInvalidateEntry(location common.AddressLocation, _ *derived.Program, _ *snapshot.ExecutionSnapshot) bool {
-	_, ok := inv.invalidated[location]
-	return inv.invalidateAll || ok
-}
-
-type meterParamOverridesInvalidator struct {
-	invalidateAll bool
-}
-
-func (inv *meterParamOverridesInvalidator) ShouldInvalidateEntries() bool {
-	return inv.invalidateAll
-}
-
-func (inv *meterParamOverridesInvalidator) ShouldInvalidateEntry(_ struct{}, _ derived.MeterParamOverrides, _ *snapshot.ExecutionSnapshot) bool {
-	return inv.invalidateAll
 }
